// src/app/invoice/model/AdobePDFInfo.ts
export class AdobePDFInfo {
  pageOrientation: string;
  expandPageWidthFitinvoice: boolean;
  topMarginValue: number;
  topMarginTo: string;
  bottomMarginValue: number;
  bottomMarginTo: string;
  fontSize: number;
  font: string;
  invoiceTitle: string;
  isPrintHeaderDoubleWindow: boolean;
  isPrintHeaderE65Window: boolean;
  isPrintHeaderDLWindow: boolean;
  isPrintHeaderC4Window: boolean;
  isIncludeShipToAddressPDFInvoice: boolean;
  isRepeatMultiPageInvoices: boolean;
  remitToAddress: string;
}